requirements.txt
attrs == 21.4.0
click == 8.0.3
jsonschema == 4.4.0
paho-mqtt == 1.6.1
psycopg == 3.0.8
pycron == 3.0.0
pyyaml == 6.0
rx == 3.2.0
schedule == 1.1.0
tzlocal == 4.1
